+++ test/interop/gohealth/main.go
@@ -1,0 +1,26 @@
+package main
+
+import (
+	"log"
+	"net"
+
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/health"
+	healthpb "google.golang.org/grpc/health/grpc_health_v1"
+	"google.golang.org/grpc/reflection"
+)
+
+func main() {
+	lis, err := net.Listen("tcp", ":50051")
+	if err != nil {
+		log.Fatal(err)
+	}
+	s := grpc.NewServer()
+	hs := health.NewServer()
+	hs.SetServingStatus("", healthpb.HealthCheckResponse_SERVING)
+	hs.SetServingStatus("livery.Test", healthpb.HealthCheckResponse_NOT_SERVING)
+	healthpb.RegisterHealthServer(s, hs)
+	reflection.Register(s)
+	log.Println("go grpc health server on :50051")
+	log.Fatal(s.Serve(lis))
+}